make build script to handle feature branches
