+++ package.json
@@ -5,11 +5,10 @@
   "main": "index.js",
   "scripts": {
     "build": "npm run build@2.0 && npm run build@3.0",
-    "build@feature-branch": "node ./scripts/feature-branch",
     "start@2.0": "(export NODE_ENV=development; concurrently --kill-others \"node_modules/hugo-bin/vendor/hugo --baseURL=/docs -d ../dist/docs -s site --watch\" \"webpack --watch --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js\")",
-    "build@2.0": "(export HUGO_VERSION=0.31.1 NODE_ENV=production; webpack --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js && node_modules/hugo-bin/vendor/hugo version && node_modules/hugo-bin/vendor/hugo --baseURL=/docs -d ../dist/docs -s site)",
+    "build@2.0": "node ./scripts/build --baseURL=/docs --output=../dist/docs",
     "start@3.0": "(export NODE_ENV=development; concurrently --kill-others \"node_modules/hugo-bin/vendor/hugo --baseURL=/beta-docs --contentDir=content@beta -d ../dist/beta-docs -s site --watch\" \"webpack --watch --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js\")",
-    "build@3.0": "(export HUGO_VERSION=0.31.1 NODE_ENV=production; webpack --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js && node_modules/hugo-bin/vendor/hugo version && node_modules/hugo-bin/vendor/hugo --baseURL=/beta-docs --contentDir=content@beta -d ../dist/beta-docs -s site)"
+    "build@3.0": "node ./scripts/build --baseURL=/beta-docs --contentDir=content@beta --output=../dist/beta-docs"
   },
   "keywords": [],
   "author": "",
@@ -25,7 +24,8 @@
     "hugo-bin": "^0.18.1",
     "laravel-mix": "^1.4.0",
     "shelljs": "^0.8.3",
-    "vue": "^2.4.2"
+    "vue": "^2.4.2",
+    "yargs": "^13.2.4"
   },
   "dependencies": {
     "bootstrap": "^4.0.0-beta.2",
